refactor(runtime): universal `ApiResourcesClient`, `ResourceResolver`
